feat: 회원가입 시 유저 정보 struct 생성
