+++ CodeFantasia/Feature/UserManager.swift
@@ -8,3 +8,38 @@ let REF_USERS = DB_REF.child("users")
 
 let STORAGE_REF = Storage.storage().reference()
 let STORAGE_PROFILE_IMAGES = STORAGE_REF.child("profile_images")
+
+struct AuthService {
+    static let shared = AuthService()
+    
+    func registerUser() {
+        guard let imageData = profileImage.jpegData(compressionQuality: 0.3) else { return }
+        let filename = NSUUID().uuidString
+        let storageRef = STORAGE_PROFILE_IMAGES.child(filename)
+        
+        storageRef.putData(imageData) { (meta, error) in
+            storageRef.downloadURL { (url, error) in
+                guard let profileImageUrl = url?.absoluteString else { return }
+                
+                Auth.auth().createUser(withEmail: email, password: password) { (result, error) in
+                    if let error = error {
+                        print("계정 등록 에러: \(error.localizedDescription)")
+                        return
+                    }
+                    
+                    guard let uid = result?.user.uid else { return }
+                    
+                    let values = ["email" : email,
+                                  "name" : name,
+                                  "profileImageUrl" : profileImageUrl]
+                    
+                    REF_USERS.child(uid).updateChildValues(values) { (error, ref) in
+                        print("실시간 데이터 베이스에 유저 정보 업데이트 성공.")
+                        print("email: \(email), name: \(name), profileImageUrl: \(profileImageUrl)")
+                    }
+                }
+                
+            }
+        }
+    }
+}
